requirements.txt
typer>=0.12.0
rich>=13.7.0
python-dotenv>=1.0.1
pydantic>=2.7.0
openai>=1.40.0
requests>=2.32.0
pdfplumber>=0.11.0
rapidfuzz>=3.9.0
numpy>=1.26.0
pytest>=8.2.0
